♻️ refactor(project structure) : move some files in a common folder
