+++ src/main.ts
@@ -1,7 +1,7 @@
 import { NestFactory } from '@nestjs/core';
-import { AppModule } from './app/app.module';
 import { ValidationPipe } from '@nestjs/common';
-import { MongoExceptionFilter } from './utils/mongoExceptionFilter/mongoExceptionFilter';
+import { AppModule } from './app/app.module';
+import { MongoExceptionFilter } from './common/filters/mongo-exception.filter';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
